+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.12.0] - 2021-09-07
+
 ### Added
 
 - Add setting to display stars in `product-rating-inline` block when the product has no reviews